Struct genomics1_beta2::ReadgroupsetMethodsBuilder

pub struct ReadgroupsetMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on readgroupset resources. It is not used directly, but through the Genomics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate genomics1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1_beta2::Genomics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `align(...)`, `call(...)`, `coveragebuckets_list(...)`, `delete(...)`, `export(...)`, `get(...)`, `import(...)`, `patch(...)`, `search(...)` and `update(...)`
// to build up your call.
let rb = hub.readgroupsets();

Methods

impl<'a, C, NC, A> ReadgroupsetMethodsBuilder<'a, C, NC, A>

fn update(&self, request: &ReadGroupSet, read_group_set_id: &str) -> ReadgroupsetUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a read group set.

fn export(&self, request: &ExportReadGroupSetsRequest) -> ReadgroupsetExportMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Exports read group sets to a BAM file in Google Cloud Storage.

Note that currently there may be some differences between exported BAM files and the original BAM file at the time of import. In particular, comments in the input file header will not be preserved, some custom tags will be converted to strings, and original reference sequence order is not necessarily preserved.

fn import(&self, request: &ImportReadGroupSetsRequest) -> ReadgroupsetImportMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates read group sets by asynchronously importing the provided information.

Note that currently comments in the input file header are not imported and some custom tags will be converted to strings, rather than preserving tag types. The caller must have WRITE permissions to the dataset.

fn patch(&self, request: &ReadGroupSet, read_group_set_id: &str) -> ReadgroupsetPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a read group set. This method supports patch semantics.

fn call(&self, request: &CallReadGroupSetsRequest) -> ReadgroupsetCallMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Calls variants on read data from existing read group sets or files from Google Cloud Storage. See the alignment and variant calling documentation for more details.

fn align(&self, request: &AlignReadGroupSetsRequest) -> ReadgroupsetAlignMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Aligns read data from existing read group sets or files from Google Cloud Storage. See the alignment and variant calling documentation for more details.

fn coveragebuckets_list(&self, read_group_set_id: &str) -> ReadgroupsetCoveragebucketListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists fixed width coverage buckets for a read group set, each of which correspond to a range of a reference sequence. Each bucket summarizes coverage information across its corresponding genomic range.

Coverage is defined as the number of reads which are aligned to a given base in the reference sequence. Coverage buckets are available at several precomputed bucket widths, enabling retrieval of various coverage 'zoom levels'. The caller must have READ permissions for the target read group set.

fn delete(&self, read_group_set_id: &str) -> ReadgroupsetDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a read group set.

fn search(&self, request: &SearchReadGroupSetsRequest) -> ReadgroupsetSearchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Searches for read group sets matching the criteria.

Implements GlobalAllianceApi.searchReadGroupSets.

fn get(&self, read_group_set_id: &str) -> ReadgroupsetGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a read group set by ID.