Struct genomics1_beta2::VariantsetMethodsBuilder

pub struct VariantsetMethodsBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on variantset resources. It is not used directly, but through the Genomics hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate genomics1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use genomics1_beta2::Genomics;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Genomics::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `delete(...)`, `export(...)`, `get(...)`, `import_variants(...)`, `merge_variants(...)`, `patch(...)`, `search(...)` and `update(...)`
// to build up your call.
let rb = hub.variantsets();

Methods

impl<'a, C, NC, A> VariantsetMethodsBuilder<'a, C, NC, A>

fn update(&self, request: &VariantSet, variant_set_id: &str) -> VariantsetUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a variant set's metadata. All other modifications are silently ignored.

fn export(&self, request: &ExportVariantSetRequest, variant_set_id: &str) -> VariantsetExportMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Exports variant set data to an external destination.

fn patch(&self, request: &VariantSet, variant_set_id: &str) -> VariantsetPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a variant set's metadata. All other modifications are silently ignored. This method supports patch semantics.

fn search(&self, request: &SearchVariantSetsRequest) -> VariantsetSearchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns a list of all variant sets matching search criteria.

Implements GlobalAllianceApi.searchVariantSets.

fn get(&self, variant_set_id: &str) -> VariantsetGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a variant set by ID.

fn merge_variants(&self, request: &MergeVariantsRequest, variant_set_id: &str) -> VariantsetMergeVariantMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Merges the given variants with existing variants. Each variant will be merged with an existing variant that matches its reference sequence, start, end, reference bases, and alternative bases. If no such variant exists, a new one will be created.

When variants are merged, the call information from the new variant is added to the existing variant, and other fields (such as key/value pairs) are discarded.

fn delete(&self, variant_set_id: &str) -> VariantsetDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes the contents of a variant set. The variant set object is not deleted.

fn import_variants(&self, request: &ImportVariantsRequest, variant_set_id: &str) -> VariantsetImportVariantMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates variant data by asynchronously importing the provided information.

The variants for import will be merged with any existing data and each other according to the behavior of mergeVariants. In particular, this means for merged VCF variants that have conflicting INFO fields, some data will be arbitrarily discarded. As a special case, for single-sample VCF files, QUAL and FILTER fields will be moved to the call level; these are sometimes interpreted in a call-specific context. Imported VCF headers are appended to the metadata already in a variant set.